+++ src/Helper/FormHelper.php
@@ -8,18 +8,18 @@ use Metabolism\WordpressBundle\Plugin\MediaPlugin as Media;
 /**
  * Class Metabolism\WordpressBundle Framework
  */
-class FormHelper {
+class Form {
 
 	/**
 	 * Get request parameter
 	 */
-	public static function get( $key, $limit_lengh=500 )
+	public static function getField( $key, $limit_lengh=500 )
 	{
 		if( isset($_FILES[$key]))
 		{
 			$upload = Media::upload($key, ['image/jpeg', 'image/gif', 'image/png', 'application/pdf', 'application/zip']);
 
-			if( isset($upload['error']))
+			if( is_wp_error($upload) )
 				return false;
 
 			return $upload['filename'];
@@ -34,87 +34,61 @@ class FormHelper {
 		}
 	}
 
-
 	/**
-	 * Serialize REQUEST data
+	 * Get whole form
 	 */
-	public static function meetRequirement( $id, $value='', $validation = [] ) {
+	public static function get($fields=[]){
 
-		//todo check for type
-		$id = str_replace('[]', '', $id);
-		return isset( $validation[$id] ) && ( !isset($validation[$id]['required']) || !empty($value) );
-	}
+		$form = [];
 
-
-	/**
-	 * Serialize REQUEST data
-	 */
-	public static function serialize( $key = 'data', $validation = [], $limit_lengh=500 ) {
-
-		$form_data = isset($_REQUEST[$key])?$_REQUEST[$key]:[];
-		$data = [];
-
-		foreach ($form_data as $field)
+		foreach ( $fields as $key )
 		{
-			$id    = trim( sanitize_text_field($field['name']) );
-			$value = substr( trim( sanitize_text_field($field['value'])), 0, $limit_lengh );
-
-			if( strpos($id, '[]') !== false )
-				$data[str_replace('[]', '', $id)][] = $value;
-			else
-				$data[$id] = $value;
-
-			if( !empty($validation) && !self::meetRequirement($id, $value, $validation) )
-				return false;
+			$form[$key] = self::getField( $key );
 		}
 
-		return $data;
+		return $form;
 	}
 
 	/**
-	 * Quickly send form
+	 * Send form
 	 */
-	public static function send($fields=[], $files=[], $subject='New message from website', $email_id='email', $delete_attachements=true){
+	public static function send($to='admin', $subject='New message from website', $fields=[], $attachements=[], $email_id='email' ){
 
-		$email = self::get( $email_id );
+		if( !in_array($email_id, $fields) )
+			$fields[] = $email_id;
 
-		if ( $email && is_email( $email ) )
+		$form = self::get($fields);
+
+		if ( is_email( $form[$email_id] ) )
 		{
+			if(!$to || $to='admin')
+				$to = get_option( 'admin_email' );
+
 			$body = $subject." :\n\n";
-
-			foreach ( $fields as $key )
-			{
-				$value = self::get( $key );
-				$body  .= ( $value ? ' - ' . $key . ' : ' . $value . "\n" : '' );
-			}
-
 			$attachments = [];
 
-			foreach ( $files as $file )
-			{
-				$file = self::get( $file );
+			foreach ( $fields as $key ) {
 
-				if ( file_exists( WP_CONTENT_DIR.$file ) )
-					$attachments[] = WP_CONTENT_DIR.$file;
+				$body .= ($form[$key] ? ' - ' . $key . ' : ' . $form[$key] . "\n" : '');
 			}
 
-			if ( wp_mail( get_option( 'admin_email' ), $subject, $body, $attachments ) )
+			foreach ( $attachements as $attachement )
 			{
-				if( $delete_attachements )
+				if ( file_exists( $attachement ) )
 				{
-					foreach ( $attachments as $file )
-						@unlink($file);
+					$attachments[] = $attachement;
 				}
-
-				return true;
 			}
+
+			if ( wp_mail( $to, $subject, $body, $attachments ) )
+				return $form;
 			else
-				return ['error' => 2, 'message' => "Sorry, the server wasn't able to complete this request"];
+				return new \WP_Error('send_mail', "The server wasn't able to send the email.");
 
 		}
 		else
 		{
-			return ['error' => 1, 'message' => "Invalid email address. Please type a valid email address."];
+			return new \WP_Error('invalid_email', "Invalid email address. Please type a valid email address.");
 		}
 	}
 }