+++ changelog.md
@@ -18,6 +18,7 @@ CHANGELOG
 - Comment entity
 - Context->addComments with replies
 - Form helper :: post comment
+- Template in wordpress.yml for all post_type
 ### Changed
 - Context->addBreadcrumb now add post/term title
 - Context->addCurrent to add term for archive